separate linux and darwin steps (#50)
Currently the release automation will fail to upload any binaries if any
part of the 'fetch envoy' step fails. This is problematic because we
currently attempt to fetch darwin binaries from Homebrew, but Homebrew
does not build older patch releases of Envoy.

As a temporary workaround, split the 'fetch envoy' step into two, and
upload to the release right after fetching. This way, even if we cannot
fetch the darwin binaries, we should still fetch and upload the linux
binaries.