fetch binary built for oldest macOS version (#51)
It looks like Homebrew builds binaries for the three latest versions of
macOS (for both amd64 and arm64). Currently it looks like the
fetch-envoy script may end up downloading any one of the three available
binaries for a specific architecture.

Let's update this script so that it will choose the binary built for the
oldest available macOS version, as that should provide the widest
compatibility. We can do this by first fetching the multi-platform
manifest, filtering by OS and architecture, and then finding the minimum
OS version.

(Note that the OS version is a text string like "macOS 12.6". Comparing
these strings may not be totally correct in general, but I think it
should give a correct result while the major version still has two
digits.)